# Cache Cldr operations to improve performance when printing timezones