Fixes #394 by adding module-level caches and memoizing functions for various operations involved in printing timezone names.

Running the same test as in #394 that prints timezone names 1000 times, we now have a drastic speedup. 

**Before**:
```
UTC without timezone: 19.212158203125ms
UTC with timezone: 11.41796875ms
NYC without timezone: 5.911865234375ms
NYC with timezone: 599.635009765625ms
```

**After**:
```
UTC without timezone: 20.73486328125ms
UTC with timezone: 10.364990234375ms
NYC without timezone: 6.211181640625ms
NYC with timezone: 6.883056640625ms
```

My main question is if it's okay to have essentially global, permanent, caches as are added in this PR? Obviously there could be memory usage concerns with these constructs. I've left a few notes inline to summarize my thoughts.