@phueper 2020-07-08
I like the PR ... i never thought of the cldr (which is actually just loading/requiring the cldr data) to make this big a difference, but it does make sense.
I would also like to get some numbers on the memory usage if that is possible.

i wonder wether this might hurt us if we ever get to fix https://github.com/js-joda/js-joda-locale/issues/93  ... i.e. when loading multiple different locale instances ... but i can't think of a reason right now ... 

So, if the review comments by @pithu are resolved and we have some memory numbers for the tests... i'd be happy to get this merged :+1: