@JKillian 2020-07-26
I just pushed up some new changes - we never use cldr instances as cache keys now, we just use locale strings, so all the caches are just plain objects now.

I also fixed the bug where `_resolveZoneIdText` was getting called instead of  `_cachedResolveZoneIdText`.

I also left a few comments above about memory consumption - everything seemed okay in my analysis (but can't say for 100% certainty that I was doing things correctly haha).